Move primitive implementations to own cpp
